refactor(compiler-cli): move signal identification function
Moves the function that identifies signals into a separate file so that it can be reused outside of extended diagnostics.
