+++ packages/compiler-cli/src/ngtsc/typecheck/extended/checks/interpolated_signal_not_invoked/index.ts
@@ -10,17 +10,10 @@ import {AST, Interpolation, PropertyRead, TmplAstNode} from '@angular/compiler';
 import ts from 'typescript';
 
 import {ErrorCode, ExtendedTemplateDiagnosticName} from '../../../../diagnostics';
-import {NgTemplateDiagnostic, SymbolKind} from '../../../api';
+import {NgTemplateDiagnostic} from '../../../api';
+import {isSignalReference} from '../../../src/symbol_util';
 import {TemplateCheckFactory, TemplateCheckWithVisitor, TemplateContext} from '../../api';
 
-/** Names of known signal functions. */
-const SIGNAL_FNS = new Set([
-  'WritableSignal',
-  'Signal',
-  'InputSignal',
-  'InputSignalWithTransform',
-  'ModelSignal',
-]);
 
 /** Names of known signal instance properties. */
 const SIGNAL_INSTANCE_PROPERTIES = new Set(['set', 'update', 'asReadonly']);
@@ -50,18 +43,6 @@ class InterpolatedSignalCheck extends
   }
 }
 
-function isSignal(symbol: ts.Symbol|undefined): boolean {
-  const declarations = symbol?.getDeclarations();
-
-  return declarations !== undefined && declarations.some(decl => {
-    const fileName = decl.getSourceFile().fileName;
-
-    return (ts.isInterfaceDeclaration(decl) || ts.isTypeAliasDeclaration(decl)) &&
-        SIGNAL_FNS.has(decl.name.text) &&
-        (fileName.includes('@angular/core') || fileName.includes('angular2/rc/packages/core'));
-  });
-}
-
 function isFunctionInstanceProperty(name: string): boolean {
   return FUNCTION_INSTANCE_PROPERTIES.has(name);
 }
@@ -76,8 +57,7 @@ function buildDiagnosticForSignal(
     Array<NgTemplateDiagnostic<ErrorCode.INTERPOLATED_SIGNAL_NOT_INVOKED>> {
   // check for `{{ mySignal }}`
   const symbol = ctx.templateTypeChecker.getSymbolOfNode(node, component);
-  if (symbol?.kind === SymbolKind.Expression &&
-      (isSignal(symbol.tsType.symbol) || isSignal(symbol.tsType.aliasSymbol))) {
+  if (symbol !== null && isSignalReference(symbol)) {
     const templateMapping =
         ctx.templateTypeChecker.getTemplateMappingAtTcbLocation(symbol.tcbLocation)!;
     const errorString = `${node.name} is a function and should be invoked: ${node.name}()`;
@@ -92,8 +72,7 @@ function buildDiagnosticForSignal(
   // `{{ mySignal.asReadonly }}` as these are the names of instance properties of Signal
   const symbolOfReceiver = ctx.templateTypeChecker.getSymbolOfNode(node.receiver, component);
   if ((isFunctionInstanceProperty(node.name) || isSignalInstanceProperty(node.name)) &&
-      symbolOfReceiver?.kind === SymbolKind.Expression &&
-      (isSignal(symbolOfReceiver.tsType.symbol) || isSignal(symbolOfReceiver.tsType.aliasSymbol))) {
+      symbolOfReceiver !== null && isSignalReference(symbolOfReceiver)) {
     const templateMapping =
         ctx.templateTypeChecker.getTemplateMappingAtTcbLocation(symbolOfReceiver.tcbLocation)!;
